+++ NEWS.md
@@ -9,6 +9,10 @@
 * `as.tbl.pccc_codes` has been deprecated since `dplyr::as.tbl` has been
   deprecated in favor of `tibble::as_tibble`
 
+## Defunct functions:
+* `test_helper` was an exported function that was used in the testing suite but
+  not needed by any end user.  It is now defunct.
+
 # Version 1.0.5
 
 Changed included datafiles to be of type data.frame to avoid some libraries trying to load data.table unnecessarily. Similarly removed dependencies on dplyr::data_frame() due to deprecation.